Correct to float values in FF

+++ backend/src/image_processing/cpp/FlatFieldor.cpp
@@ -94,10 +94,10 @@ void FlatFieldor::execute(CallBackFunction func, btrgb::ArtObject* images) {
 */
 void::FlatFieldor::wCalc(int base, int rings, int patX, int patY, double yRef, btrgb::Image* a1, btrgb::Image* a2, btrgb::Image* wh1, btrgb::Image* wh2){
     //Setting values for the For Loop going over one channel, channel 2
-    int art1Total = 0;
-    int white1Total = 0;
-    int art2Total = 0;
-    int white2Total = 0;
+    float art1Total = 0;
+    float white1Total = 0;
+    float art2Total = 0;
+    float white2Total = 0;
     int loops = 0;
     int xOff, yOff, currRow, currCol;
 
@@ -140,7 +140,7 @@ void::FlatFieldor::pixelOperation(int h, int w, int c, btrgb::Image* a1, btrgb::
     //For loop is for every pixel in the image, and gets a corrisponding pixel from white and dark images
     //Every Channel value for each pixel needs to be adjusted based on the w for that group of images
     int currRow, currCol, ch;
-    int wPix, dPix, aPix, newPixel;
+    float wPix, dPix, aPix, newPixel;
     for (currRow = 0; currRow < h; currRow++) {
         for (currCol = 0; currCol < w; currCol++) {
             for (ch = 0; ch < c; ch++) {
